requirements.txt
beautifulsoup4>=4.6.0
Django>=1.10,<1.11
django-admin-sortable2>=0.6.10
django-allauth>=0.31.0
django-autocomplete-light>=3.2.0
django-braces>=1.8.1
django-choices>=1.5.0
django-ckeditor>=5.0.0
django-ckeditor-filebrowser-filer>=0.2.1
django-cms>=3.4.0
django-colorful>=1.2
django-crispy-forms>=1.6.0
django-daterange-filter>=1.2.0
django-dynamic-preferences>=1.3
django-easy-pdf>=0.1.0
django-filer>=1.2.7
django-ical>=1.4
django-imagekit>=3.3
django-multiselectfield>=0.1.5
django-polymorphic>=1.0.2
django-sekizai>=0.10.0
djangocms-admin-style>=1.2.6.2
djangocms-forms>=0.2.5
djangocms-text-ckeditor>=3.4.0
easy-thumbnails>=2.3
huey>=1.3.1
icalendar>=3.9.0
intervaltree>=2.1.0
paypalrestsdk>=1.12.0
peewee>=2.10.1
Pillow>=3.4.2
python-dateutil>=2.4.1
pytz>=2017.2
redis>=2.10.6
requests>=2.6.0
six>=1.10.0
stripe>=1.62.0
unicodecsv>=0.14.1
